requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
pydantic==2.3.0
python-dotenv==1.0.0
email-validator==2.0.0
PyMySQL==1.1.0
cryptography==41.0.4